requirements.txt
Django==4.0.3
djangorestframework==3.12.4
django-widget-tweaks==1.4.8
image==1.5.33
Pillow==9.1.0
pytest==7.1.1
pytz
asgiref==3.5.0
django-object-actions==4.0.0


psycopg2==2.9.5
django-phonenumber-field==7.0.0
phonenumbers==8.12.57


